% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-variables.R
\name{basic_variables}
\alias{basic_variables}
\title{Basic variables}
\usage{
basic_variables(dt)
}
\arguments{
\item{dt}{input data.table}
}
\value{
A list with the following elements is returned
 \item{output_measurements}{data.table with}
 \itemize{
     \item{\code{subject}} {subject ID}
     \item{\code{measurement_x}} {results of measurement with X}
     \item{\code{measurement_y}} {results of measurement with Y}
     \item{\code{d_ij}} {difference of each pair of measurement}
     \item{\code{m_ij}} {mean of each pair of measurement}
 }
 \item{output_subjects}{data.table with}
 \itemize{
     \item{\code{subject}} {subject ID}
     \item{\code{m_i}} {number of measurements of each subject}
     \item{\code{d_i}} {mean of differences for each subject}
     \item{\code{var_d_i}} {variance of d_i for each subject}
 }
  \item{n}{number of subjects}
  \item{n_obs}{number of observations}
  \item{d}{mean of all differences}
  \item{d_a}{alternative mean of all differences}
  \item{mean_x}{mean of all measurements with X}
  \item{mean_x_a}{alternative mean of all measurements with X}
  \item{mean_y}{mean of all measurements with Y}
  \item{mean_y_a}{alternative mean of all measurements with Y}
  \item{rep_coeff}{repeatability coefficients}
}
\description{
\code{basic_variables} returns basic variables of an input
data.table containing three columns (subject, measurement_x and measurement_y).
}
\note{
'Alternative' means are calculated with mean for each subject
and then the overall mean of all subjects (every subject has the same impact
on the overall mean).
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
