% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtr.R
\docType{package}
\name{blandxtr-package}
\alias{blandxtr-package}
\alias{_PACKAGE}
\title{\code{blandxtr}: A package for advanced Bland Altman analysis}
\description{
blandxtr realises comparison of two methods for 
    repeated measurements with changing true value of the measured
    quantity based on advanced Bland Altman analysis as proposed by
    E Olofsen et al.in 2015 <doi:10.1007/s10877-014-9577-3>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ikoenemund/blandxtr.git}
  \item Report bugs at \url{https://github.com/ikoenemund/blandxtr/issues}
}

}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
