% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtr.R
\name{blandxtr}
\alias{blandxtr}
\title{Main method for blandxtr}
\usage{
blandxtr(input_dt, bt, bias_mod, alpha, beta)
}
\arguments{
\item{input_dt}{data.table with input dataset}

\item{bt}{number of bootstrap samples (no bootstrapping if bt <= 0)}

\item{bias_mod}{set TRUE for modified calculation of bias (small wsv) and
its variance, set FALSE for standard calculation of bias (small bsv) and
its variance}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list (blandxtr S3 object) containing the return values of all used
functions (includes tables and ggplot2 objects).
}
\description{
\code{blandxtr} performs modified Bland Altman-analysis as
proposed by Olofsen et al. (2015). Uses the following functions from
blandxtr-package: \code{basic_variables}, \code{var_tvv}, \code{loa},
\code{var_loa}, \code{ci_loa_ba},  \code{ci_loa_bt} and \code{ci_loa_mover}.
}
\note{
\code{bias_mod} is automatically set TRUE for
different number of measurements in each subject (unbalanced case)

"_mod" labels results based on modified true value varies-method

Bootstrapping affects runtime severely. Set bt<=0
if you want to skip bootstrapping.
}
\author{
Inga Koenemund \email{inga.koenemund(at)web.de}
}
