% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtr.ci.R
\name{blandxtr_ci}
\alias{blandxtr_ci}
\title{Confidence intervals for LoA (3 methods)}
\usage{
blandxtr_ci(bt, input_dt, biasMod, bv, var_tvv, loa, loa_mod, var_loa,
  var_loa_mod, alpha, beta)
}
\arguments{
\item{bt}{number of bootstrap samples}

\item{input_dt}{data.table with input dataset}

\item{biasMod}{set TRUE for modified calculation of bias (small wsv),
set FALSE for standard calculation of bias (small bsv)}

\item{bv}{result of \code{basicVariables.R}}

\item{var_tvv}{result of \code{var_tvv.R}}

\item{loa}{result of \code{loa.R}}

\item{loa_mod}{result of \code{loa.R} based on modified tvv}

\item{var_loa}{result of \code{var.loa.R}}

\item{var_loa_mod}{result of \code{var.loa.R} based on modified tvv}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list with confidence intervals of the three methods
}
\description{
\code{blandxtr_ci} returns confidence intervals (CI)
for limits of agreement (LoA) based on three different methods
(Bland and Altman (1999), MOVER and parametric bootstrap-t)
}
\note{
Input can easily generated by previously running
\code{blandxtrMain.pre}
}
\seealso{
[blandxtrMain_pre()]
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
