% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtr.results.plot.R
\name{blandxtr_results_plot}
\alias{blandxtr_results_plot}
\title{Visualization (plot) of the results of modified Bland Altman-analysis}
\usage{
blandxtr_results_plot(res)
}
\arguments{
\item{res}{list with results from \code{blandxtrMain}}
}
\value{
Plot showing analysis results

Plot showing modified analysis results

QQ-plot of individual means

QQ-plot of individual residuals

Plot of residuals vs mean

Plot of residuals vs ID
}
\description{
\code{blandxtr.results.plot} visualizes (plot) the results of
modified Bland Altman-analysis performed with \code{blandxtrMain}
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
