% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tables.R
\name{blandxtr_results_table}
\alias{blandxtr_results_table}
\title{Tables with results of modified Bland Altman-analysis}
\usage{
blandxtr_results_table(res, bt, bias_mod, alpha, beta)
}
\arguments{
\item{res}{list with results from \code{blandxtrMain}}

\item{bt}{number of bootstrap samples (no bootstrapping if bt <= 0)}

\item{bias_mod}{set TRUE for modified calculation of bias (small wsv) and
its variance, set FALSE for standard calculation of bias (small bsv) and
its variance}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
Table with analysis results

Table with modified analysis results

Table with repeatability coefficients

Table with individual residuals

Table with residuals
}
\description{
\code{blandxtr.results.table} summarizes the results of
modified Bland Altman-analysis performed with \code{blandxtrMain} as
LaTeX-tables.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
