% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtrMain.pre.R
\name{blandxtrMain_pre}
\alias{blandxtrMain_pre}
\title{Helper function for main method for blandxtr}
\usage{
blandxtrMain_pre(input_dt, bt, biasMod)
}
\arguments{
\item{input_dt}{data.table with input dataset}

\item{bt}{number of bootstrap samples}

\item{biasMod}{set TRUE for modified calculation of bias (small wsv),
set FALSE for standard calculation of bias (small bsv)}
}
\value{
A list containing the return values of all used functions.
}
\description{
\code{blandxtrMain_pre} performs modified Bland Altman-analysis
as proposed by Olofsen et al. (2015) without calculation of
confidence intervals. Helper function for \code{blandxtr.Main} which performs
the whole analysis.
}
\note{
\code{biasMod} is automatically set TRUE for
different number of measurements in each subject (unbalanced case)

"_mod" labels results based on modified true value varies-method
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
