% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.loa.bt.R
\name{calc_ci_loa_bt}
\alias{calc_ci_loa_bt}
\title{95\%-confidence intervals for LoA (parametric bootsrap-t)}
\usage{
calc_ci_loa_bt(bt, input_dt, biasMod, loa_l, loa_u, var_loa)
}
\arguments{
\item{bt}{number of bootstrap samples}

\item{input_dt}{data.table with input dataset}

\item{biasMod}{set TRUE for modified calculation of bias (small wsv),
set FALSE for standard calculation of bias (small bsv)}

\item{loa_l}{lower limit of agreement}

\item{loa_u}{upper limit of agreement}

\item{var_loa}{variance of limits of agreement}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{ci_l_loa_l_bt}} {lower limit of 95\%-CI for lower LoA}
 \item{\code{ci_u_loa_l_bt}} {upper limit of 95\%-CI for lower LoA}
 \item{\code{ci_l_loa_u_bt}} {lower limit of 95\%-CI for upper LoA}
 \item{\code{ci_u_loa_u_bt}} {upper limit of 95\%-CI for upper LoA}
}
}
\description{
\code{calc_ci_loa_bt} returns 95\%-confidence intervals
(95\%-CI) for limits of agreement (LoA). Calculation is based on
parametric bootstrap-t.
}
\note{
\code{biasMod} is automatically set TRUE for
different number of measurements in each subject (unbalanced case)

"_mod" labels results based on modified true value varies-method
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
