% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.tvv.R
\name{calc_var_tvv}
\alias{calc_var_tvv}
\title{Analysis of variance}
\usage{
calc_var_tvv(n, n_obs, d, d_a, outputSubjects, outputMeasurements)
}
\arguments{
\item{n}{number of subjects}

\item{n_obs}{number of measurements}

\item{d}{mean of all differences}

\item{d_a}{modified mean of all differences}

\item{outputSubjects}{data.table containing subject ID and
number of measurements of each subject (m_i)}

\item{outputMeasurements}{data.table containing}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{wsv}} {within-subjects variance}
 \item{\code{bsv}} {between-subjects variance}
 \item{\code{bsv_mod}} {modified between-subjects variance}
 \item{\code{wsv_mod}} {modified within-subjects variance}
 \item{\code{var_d}} {variance of mean of all differences}
 \item{\code{sd_d}} {standard deviation of all differences}
 \item{\code{var_var_d}} {variance of the variance of mean
 of all differences}
 \item{\code{var_d_mod}} {modified variance of mean of all differences}
 \item{\code{sd_d_mod}} {modified standard deviation of all differences}
 \item{\code{var_var_d_mod}} {modified variance of the variance of
 mean of all differences}
}
}
\description{
\code{calc_var_tvv} executes analysis of variance for
quantities measured with true value varying
(includes standard and modified version)
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
