% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-loa-bt.R
\name{ci_loa_bt}
\alias{ci_loa_bt}
\title{Confidence intervals for limits of agreement (using parametric bootsrap-t)}
\usage{
ci_loa_bt(bt, input_dt, bias_alt, loa_l, loa_u, var_loa, alpha, beta)
}
\arguments{
\item{bt}{number of bootstrap samples}

\item{input_dt}{data.table with input dataset}

\item{bias_alt}{set TRUE for alternative calculation of bias (small
within-subject variance) and its variance, set FALSE for standard calculation
of bias (small between-subjects variance) and its variance}

\item{loa_l}{lower limit of agreement}

\item{loa_u}{upper limit of agreement}

\item{var_loa}{variance of limits of agreement}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{ci_l_loa_l_bt}} {lower limit of 95\%-CI for lower LoA}
 \item{\code{ci_u_loa_l_bt}} {upper limit of 95\%-CI for lower LoA}
 \item{\code{ci_l_loa_u_bt}} {lower limit of 95\%-CI for upper LoA}
 \item{\code{ci_u_loa_u_bt}} {upper limit of 95\%-CI for upper LoA}
}
}
\description{
\code{ci_loa_bt} returns confidence intervals
(CI) for limits of agreement (LoA). Calculation is based on
parametric bootstrap-t.
}
\note{
For further information please see:
\itemize{
 \item Efron, B. and Tibshirani, R. J., An introduction to the bootstrap.
 Monographs on statistics and applied probability, 57, 1993.
 \item Carpenter, J. and Bithell, J., Bootstrap confidence intervals:
 when, which, what? A practical guide for medical statisticians.
 Statistics in medicine, 19 (9), 2000.
 }
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
