% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-loa-mover.R
\name{ci_loa_mover}
\alias{ci_loa_mover}
\title{Confidence intervals for limits of agreement (using MOVER method)}
\usage{
ci_loa_mover(n, n_obs, lambda_mod, output_subjects, mssi_mod, wsv,
  var_d_mod, loa_l, loa_u, alpha, beta)
}
\arguments{
\item{n}{number of subjects}

\item{n_obs}{number of observations}

\item{lambda_mod}{helper variable for calculation of modified
between-subjects variance, see Olofsen et al. 2015}

\item{output_subjects}{data.table containing subject ID and
number of measurements of each subject (m_i)}

\item{mssi_mod}{mssi (modified calculation)}

\item{wsv}{within-subject variance (unmodified)}

\item{var_d_mod}{{modified variance of mean of all differences}}

\item{loa_l}{lower limit of agreement}

\item{loa_u}{upper limit of agreement}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{ci_l_loa_l_mover}} {lower limit of 95\%-CI for lower LoA}
 \item{\code{ci_u_loa_l_mover}} {upper limit of 95\%-CI for lower LoA}
 \item{\code{ci_l_loa_u_mover}} {lower limit of 95\%-CI for upper LoA}
 \item{\code{ci_u_loa_u_mover}} {upper limit of 95\%-CI for upper LoA}
}
}
\description{
\code{ci_loa_mover} returns confidence intervals
(CI) for limits of agreement (LoA) based on MOVER-method
(Methods of Variance Estimates Recovery) proposed
by Zou (2013).
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
