% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tables.R
\name{generate_tables}
\alias{generate_tables}
\title{Tables with results of advanced Bland Altman-analysis}
\usage{
generate_tables(res)
}
\arguments{
\item{res}{list with results from \code{blandxtr}}
}
\value{
Table with analysis results

Table with modified analysis results

Table with parameters for calculation of repeatability coefficients

Table with individual residuals

Table with residuals

Table with input parameters

Table with input data
}
\description{
\code{genrate_tables} summarizes the results of
advanced Bland Altman-analysis performed with \code{blandxtr} as
tables.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
