% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtr.R
\name{is.blandxtr}
\alias{is.blandxtr}
\title{Reports whether x is a blandxtr object}
\usage{
is.blandxtr(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a blandxtr object
}
\keyword{internal}
