% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loa.R
\name{loa}
\alias{loa}
\title{Limits of agreement (LoA)}
\usage{
loa(d, sd_d, beta)
}
\arguments{
\item{d}{mean of all differences}

\item{sd_d}{standard deviation of d}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{loa_l}} {lower limit of agreement}
 \item{\code{loa_u}} {upper limit of agreement}
}
}
\description{
\code{loa} returns limits of agreement (LoA).
Calculation is based on a method proposed by Bland and Altman (1999).
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
