% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-RData.R
\name{loadRData}
\alias{loadRData}
\title{Load RData-files}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{name of the rData-file}
}
\value{
A data.frame containing data from fileName.
}
\description{
\code{loadRData} returns a data.frame from a RData-file.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
