% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ci.R
\name{main_ci}
\alias{main_ci}
\title{Confidence intervals for LoA (3 methods)}
\usage{
main_ci(bt, input_dt, bias_alt, bv, var_tvv, loa, loa_mod, var_loa,
  var_loa_mod, alpha, beta)
}
\arguments{
\item{bt}{number of bootstrap samples}

\item{input_dt}{data.table with input dataset}

\item{bias_alt}{set TRUE for alternative calculation of bias (small
within-subject variance) and its variance, set FALSE for standard calculation
of bias (small between-subjects variance) and its variance}

\item{bv}{result of \code{basic_variables.R}}

\item{var_tvv}{result of \code{var_tvv.R}}

\item{loa}{result of \code{loa.R}}

\item{loa_mod}{result of \code{loa.R} based on modified tvv}

\item{var_loa}{result of \code{var_loa.R}}

\item{var_loa_mod}{result of \code{var_loa.R} based on modified tvv}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list with confidence intervals of the three methods
}
\description{
\code{main_ci} returns confidence intervals (CI)
for limits of agreement (LoA) based on three different methods
(Bland and Altman (1999), MOVER and parametric bootstrap-t)
}
\note{
Input can easily generated by previously running
\code{blandxtrMain.pre}
}
\seealso{
[main_pre()]
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
