% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-pre.R
\name{main_pre}
\alias{main_pre}
\title{Helper function for main method for blandxtr}
\usage{
main_pre(input_dt, bt, bias_alt, beta)
}
\arguments{
\item{input_dt}{data.table with input dataset}

\item{bt}{number of bootstrap samples}

\item{bias_alt}{set TRUE for alternative calculation of bias (small
within-subject variance) and its variance, set FALSE for standard calculation
of bias (small between-subjects variance) and its variance}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
A list containing the return values of all used functions.
}
\description{
\code{main_pre} performs advanced Bland Altman-analysis
as proposed by Olofsen et al. (2015) without calculation of
confidence intervals. Helper function for \code{blandxtr} which performs
the whole analysis.
}
\note{
"_mod" labels results based on modified analysis of variance
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
