% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olofsen}
\alias{olofsen}
\title{Results of repeated measurements}
\format{A data frame with 300 rows and 3 variables}
\source{
\url{https://sec.lumc.nl/method_agreement_analysis/sim.tx}
}
\usage{
olofsen
}
\description{
Example dataset of Olofsen et al. containing results of
15 repeated measurements (second and third column) with
two different tools in 20 individuals (first column).
The variables are as follows:
}
\details{
\describe{
  \item{subject}{subject ID}
  \item{measurement_x}{results of measuring instrument X}
  \item{measurement_y}{results of measuring instrument Y}
}
}
\keyword{datasets}
