% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blandxtr}
\alias{print.blandxtr}
\title{Print main results of modified Bland Altman-analysis}
\usage{
\method{print}{blandxtr}(x, ...)
}
\arguments{
\item{x}{list with results from \code{blandxtr}}

\item{...}{other arguments not used by this method}
}
\description{
\code{print.blandxtr} prints main results of
modified Bland Altman-analysis performed with \code{blandxtr} to the console
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
