% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_data}
\alias{read_data}
\title{Read input dataset for blandxtr}
\usage{
read_data(path)
}
\arguments{
\item{path}{path where input dataset is stored (as csv)}
}
\value{
A data.table containing input data for further analysis
}
\description{
\code{blandxtr_readData} reads input dataset (as csv) from
given path and returns a data.table for further analysis
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
