% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create report for modified Bland Altman-analysis}
\usage{
report(tab, fig, output_format, output_dir)
}
\arguments{
\item{tab}{A list of tables to be displayed (generated with
\code{generate-tables}-function from \code{blandxtr}-package}

\item{fig}{A list of plots to be displayed (generated with
\code{plot}-function from \code{blandxtr}-package (\code{plot.blandxtr})}

\item{output_format}{'html_document' (default), 'pdf_document' or
'word_document'}

\item{output_dir}{String specifying the directory for storing the output file}
}
\value{
A report of modified Bland Altman-analysis' results.
}
\description{
\code{report} returns a report of results (tables and figures)
of modified Bland Altman-analysis as html-, pdf- or word-document.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
