% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-loa.R
\name{var_loa}
\alias{var_loa}
\title{Variance of limits of agreement}
\usage{
var_loa(n, n_obs, bsv, wsv, output_subjects, var_var_d, bias_alt, beta)
}
\arguments{
\item{n}{number of subjects}

\item{n_obs}{number of observations}

\item{bsv}{between-subject variance}

\item{wsv}{within-subject variance}

\item{output_subjects}{data.table containing subject ID and
number of measurements of each subject (m_i)}

\item{var_var_d}{variance of the variance of mean of all differences}

\item{bias_alt}{set TRUE for alternative calculation of bias (small
within-subject variance) and its variance, set FALSE for standard calculation
of bias (small between-subjects variance) and its variance}

\item{beta}{for 100*(1-beta)\%-confidence interval around bias}
}
\value{
\code{var_loa} variance of limits of agreement
}
\description{
\code{calc_var_loa} returns variance of
limits of agreement (LoA) based on a method proposed
by Bland and Altman (1999).
}
\note{
function calculates other variables as well (but does not return)
\itemize{
 \item{\code{ev_var_d}} {expected value of variance of all differnces}
 \item{\code{var_d}} {variance of mean of all differences}
 \item{\code{var_d_mod}} {modified variance of mean of all differences}
 \item{\code{var_sd_d}} {variance of the standard deviation of the differences}
}
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
