% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-tvv.R
\name{var_tvv}
\alias{var_tvv}
\title{Analysis of variance}
\usage{
var_tvv(n, n_obs, d, d_a, bias_alt, output_subjects, output_measurements)
}
\arguments{
\item{n}{number of subjects}

\item{n_obs}{number of observations}

\item{d}{mean of all differences}

\item{d_a}{modified mean of all differences}

\item{bias_alt}{set TRUE for alternative calculation of bias (small
within-subject variance) and its variance, set FALSE for standard calculation
of bias (small between-subjects variance) and its variance}

\item{output_subjects}{data.table containing subject ID and
number of measurements of each subject (m_i)}

\item{output_measurements}{data.table from \code{basic_variables}-function}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{wsv}} {within-subject variance}
 \item{\code{bsv}} {between-subjects variance}
 \item{\code{bsv_mod}} {modified between-subjects variance}
 \item{\code{wsv_mod}} {modified within-subject variance}
 \item{\code{var_d}} {variance of mean of all differences}
 \item{\code{sd_d}} {standard deviation of all differences}
 \item{\code{var_var_d}} {variance of the variance of mean
 of all differences}
 \item{\code{var_d_mod}} {modified variance of mean of all differences}
 \item{\code{sd_d_mod}} {modified standard deviation of all differences}
 \item{\code{var_var_d_mod}} {modified variance of the variance of
 mean of all differences}
 \item{\code{tau}} {if \code{tau} < 1/3 use unmodified analysis of variance,
 if \code{tau} > 1/3 look at \code{tau_mod}}
 \item{\code{tau_mod}} {look at \code{tau} first, if \code{tau_mod} < 1/3 use
 unmodified analysis of variance, if \code{tau_mod} > 1/3 use
 modified analysis of variance}
 \item{\code{se_sd_d}} {standard error of sd of all differences}
 \item{\code{se_sd_d_mod}} {standard error of modified sd of all differences}
 \item{\code{se_bsv}} {standard error of between-subjects variance}
 \item{\code{se_bsv_mod}} {standard error of modified between-subjects
 variance}
 \item{\code{se_wsv}} {standard error of within-subject variance}
 \item{\code{se_wsv_mod}} {standard error of modified within-subject
 variance}
 \item{\code{lambda}} {helper variable for calculation of between-subjects
 variance, see Olofsen et al. 2015}
 \item{\code{lambda_mod}} {helper variable for calculation of modified
 between-subjects variance, see Olofsen et al. 2015}
}
}
\description{
\code{var_tvv} executes analysis of variance for
quantities measured with true value varying
(includes standard and modified version)
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
