% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{basicVariables}
\alias{basicVariables}
\title{Basic variables}
\usage{
basicVariables(dt)
}
\arguments{
\item{dt}{input data.table}
}
\value{
A list with the following elements is returned
 \item{outputMeasurements}{data.table with}
 \itemize{
     \item{\code{subject}} {subject ID}
     \item{\code{measurementX}} {results of measurement with X}
     \item{\code{measurementY}} {results of measurement with Y}
     \item{\code{d_ij}} {difference of each pair of measurement}
     \item{\code{m_ij}} {mean of each pair of measurement}
 }
 \item{outputSubjects}{data.table with}
 \itemize{
     \item{\code{subject}} {subject ID}
     \item{\code{m_i}} {number of measurements of each subject}
     \item{\code{d_i}} {mean of differences for each subject}
     \item{\code{var_d_i}} {variance of d_i for each subject}
 }
  \item{n}{number of subjects}
  \item{n_obs}{number of measurements}
  \item{d}{mean of all differences}
  \item{d_a}{modified mean of all differences}
}
\description{
\code{basicVariables} returns basic variables of an input
data.table containing three columns (subject, measurementX and measurementY).
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
