% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtrMain.R
\name{blandxtrMain}
\alias{blandxtrMain}
\title{Main method for blandxtr}
\usage{
blandxtrMain(input_dt, bt, biasMod)
}
\arguments{
\item{input_dt}{data.table with input dataset}

\item{bt}{number of bootstrap samples (no bootstrapping if bt <= 0)}

\item{biasMod}{set TRUE for modified calculation of bias (small wsv),
set FALSE for standard calculation of bias (small bsv)}
}
\value{
A list containing the return values of all used functions.
}
\description{
\code{blandxtrMain} performs modified Bland Altman-analysis as
proposed by Olofsen et al. (2015). Uses the following functions from
blandxtr-package: \code{basicVariables}, \code{var.tvv}, \code{loa},
\code{var.loa}, \code{ci.loa.ba} and \code{ci.loa.mover}.
}
\note{
\code{biasMod} is automatically set TRUE for
different number of measurements in each subject (unbalanced case)

"_mod" labels results based on modified true value varies-method

Bootstrapping affects runtime severely. Set bt<=0
if you want to skip bootstrapping.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
