% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-loa-ba.R
\name{calc_ci_loa_ba}
\alias{calc_ci_loa_ba}
\title{Confidence intervals for LoA (Bland Altman)}
\usage{
calc_ci_loa_ba(loa_l, loa_u, var_loa, alpha)
}
\arguments{
\item{loa_l}{lower limit of agreement}

\item{loa_u}{upper limit of agreement}

\item{var_loa}{variance of limits of agreement}

\item{alpha}{for 100*(1-alpha)\%-confidence interval around LoA}
}
\value{
A list with the following elements is returned
\itemize{
 \item{\code{ci_l_loa_l_ba}} {lower limit of 95\%-CI for lower LoA}
 \item{\code{ci_u_loa_l_ba}} {upper limit of 95\%-CI for lower LoA}
 \item{\code{ci_l_loa_u_ba}} {lower limit of 95\%-CI for upper LoA}
 \item{\code{ci_u_loa_u_ba}} {upper limit of 95\%-CI for upper LoA}
}
}
\description{
\code{calc_ci_loa_ba} returns confidence intervals (CI)
for limits of agreement (LoA) based on a method proposed
by Bland and Altman (1999).
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
