% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tables.R
\name{generate_tables}
\alias{generate_tables}
\title{Tables with results of modified Bland Altman-analysis}
\usage{
generate_tables(res)
}
\arguments{
\item{res}{list with results from \code{blandxtr}}
}
\value{
Table with analysis results

Table with modified analysis results

Table with repeatability coefficients

Table with individual residuals

Table with residuals
}
\description{
\code{blandxtr.results.table} summarizes the results of
modified Bland Altman-analysis performed with \code{blandxtrMain} as
LaTeX-tables.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
