% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandxtr.results.plot.R
\name{plot.blandxtr}
\alias{plot.blandxtr}
\title{Visualization (plot) of the results of modified Bland Altman-analysis}
\usage{
\method{plot}{blandxtr}(x, ...)
}
\arguments{
\item{x}{list with results from \code{blandxtrMain}}

\item{...}{other arguments not used by this method}
}
\value{
Plot showing analysis results

Plot showing modified analysis results

QQ-plot of individual means

QQ-plot of individual residuals

Plot of residuals vs mean

Plot of residuals vs ID
}
\description{
\code{plot.blandxtr} visualizes (plot) the results of
modified Bland Altman-analysis performed with \code{blandxtrMain}
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
