% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Preparation of data for Bland Altman analysis}
\usage{
prepare_data(input_dt)
}
\arguments{
\item{input_dt}{data.table with input dataset}
}
\value{
Prepared input data as data.table
}
\description{
\code{blandxtr.prepareData} checks input data.table for
common errors (missing columns that are required, wrong type of input data)
and reduces it to the columns necessary for analysis with \code{blandxtrMain}.
}
\note{
Rows with missing values will be removed from analysis.
}
\author{
Inga Koenemund \email{inga.koenemund@web.de}
}
